Add files via upload
This is a Building Integrated Photo Voltaic (BIPV) project, designed for 10 major cities of china using the record year of 2019 and 2020 to find trends, seasonal information, necessities, challenges, benefits and the energy efficiency by integrating solar panels to the skyscrapers. Machine Learning, Deep Learning and Clustering analysis was applied to combat the urban air pollution to ensure the benefit of the economy and the maximum energy production months, seasons and hours.